+++ service/auditmanager/errors.go
@@ -11,7 +11,8 @@ const (
 	// ErrCodeAccessDeniedException for service response error code
 	// "AccessDeniedException".
 	//
-	// You do not have sufficient access to perform this action.
+	// Your account is not registered with AWS Audit Manager. Check the delegated
+	// administrator setup on the AWS Audit Manager settings page, and try again.
 	ErrCodeAccessDeniedException = "AccessDeniedException"
 
 	// ErrCodeInternalServerException for service response error code